Update Data Preprocessing
n this code, first, the sample audio data is defined as a numpy array. Then, the data is converted to decibel values using 20 * np.log10(), amplitude values are obtained by multiplying the audio data by 2, and wavelength values are calculated using the operation 1 / audio_data. The converted values are then printed to the console.

This code can be used to process audio data in a shorter and simpler way.